+++ CHANGELOG.md
@@ -12,6 +12,10 @@ Version listings include the following sections, if applicable:
 
 See also <https://github.com/xgi/castero/releases>.
 
+## Unreleased
+**Added**
+* You can now delete a downloaded episode by pressing the save key again.
+
 ## 0.2.1 - 2018-04-22
 **Added**
 * Automatically migrate users' config file when the client updates and adds or